Revert "seeing if keyslots are ever valid"
This reverts commit 26d66d5a8dc804aa5c28f3c84ca1efaaf278765c.

+++ tests/fuzz/oss-fuzz-build.sh
@@ -133,7 +133,9 @@ make -j fuzz-targets
 
 for fuzzer in $ENABLED_FUZZERS; do
     cp tests/fuzz/$fuzzer $OUT
-    cp $SRC/cryptsetup_fuzzing/${fuzzer}_seed_corpus.zip $OUT
+    if [ -e $SRC/cryptsetup_fuzzing/${fuzzer}_seed_corpus.zip ]; then
+      cp $SRC/cryptsetup_fuzzing/${fuzzer}_seed_corpus.zip $OUT
+    fi
 
     # optionally copy the dictionary if it exists
     if [ -e tests/fuzz/${fuzzer}.dict ]; then
